MAJ 17/11/2021
Refractor repositories logic

+++ app/src/main/java/com/openclassrooms/realestatemanager/activity/addoredit/fragment/photos/AOEPhotoViewState.kt
@@ -3,5 +3,5 @@ package com.openclassrooms.realestatemanager.activity.addoredit.fragment.photos
 import com.openclassrooms.realestatemanager.model.Photo
 
 
-data class AOEPhotoViewState (val id: Int?, val listPhoto: List<Photo>?)
+data class AOEPhotoViewState (val id: Long?, val listPhoto: List<Photo>?)
